moves sharede tests configs and conftest
